Move some Agent Core owned packages to `internal` folders
