+++ pkg/metadata/internal/gohai/gohai.go
