+++ pkg/metadata/internal/gohai/gohai_test.go
