+++ pkg/metadata/internal/gohai/payload.go
