+++ pkg/metadata/internal/resources/resources_test.go
