+++ pkg/metadata/internal/resources/resources_windows.go
