+++ pkg/metadata/v5/payload_common.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/metadata/common"
 	"github.com/DataDog/datadog-agent/pkg/metadata/host"
-	"github.com/DataDog/datadog-agent/pkg/metadata/resources"
+	"github.com/DataDog/datadog-agent/pkg/metadata/internal/resources"
 )
 
 // CommonPayload wraps Payload from the common package
