+++ pkg/metadata/v5/payload_test.go
@@ -12,7 +12,7 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/metadata/gohai"
+	"github.com/DataDog/datadog-agent/pkg/metadata/internal/gohai"
 	"github.com/stretchr/testify/require"
 )
