+++ pkg/serializer/internal/stream/compressor.go
