+++ pkg/serializer/internal/stream/compressor_test.go
