+++ pkg/serializer/internal/stream/json_payload_builder.go
